+++ ltdconveyor/cli/main.py
@@ -18,8 +18,8 @@ CONTEXT_SETTINGS = dict(help_option_names=['-h', '--help'])
 @click.option(
     '--log-level', 'log_level',
     type=click.Choice(['warning', 'info', 'debug']),
-    default='warning',
-    help='Logging level (for first-party messages). Default: `warning`.'
+    default='info',
+    help='Logging level (for first-party messages). Default: `info`.'
 )
 @click.option(
     '--host', 'keeper_hostname',
@@ -46,8 +46,12 @@ def main(ctx, log_level, keeper_hostname, username, password):
     Use ltd to upload new site builds, and to work with the LTD Keeper API.
     """
     ch = logging.StreamHandler()
-    formatter = logging.Formatter(
-        '%(asctime)s %(levelname)8s %(name)s | %(message)s')
+    if log_level in ('warning', 'info'):
+        formatter = logging.Formatter(
+            '%(asctime)s %(levelname)8s | %(message)s')
+    else:
+        formatter = logging.Formatter(
+            '%(asctime)s %(levelname)8s %(name)s | %(message)s')
     ch.setFormatter(formatter)
 
     logger = logging.getLogger('ltdconveyor')